+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 authclient extension Change Log
 2.1.3 under development
 -----------------------
 
-- no changes in this release.
+- Bug #152: Fixed `\yii\authclient\OAuth1::fetchRequestToken()` skips formatting for `yii\httpclient\Request` (klimov-paul)
 
 
 2.1.2 February 15, 2017