+++ CHANGELOG.md
@@ -10,8 +10,8 @@ Yii Framework 2 authclient extension Change Log
 - Enh #156: Added `\yii\authclient\signature\RsaSha` and `\yii\authclient\signature\HmacSha` supporting general 'SHAwithRSA' and 'HMAC SHA' signature methods (klimov-paul)
 - Enh #157: Added `\yii\authclient\OAuth2::authenticateUserJwt()` supporting authentication via JSON Web Token (JWT) (klimov-paul)
 - Enh #163: Added support for exchanging access token at `yii\authclient\clients\Facebook` (klimov-paul)
-- Enh: Added support for client-specific access tokens at `yii\authclient\clients\Facebook` (klimov-paul)
-- Chg: `yii\authclient\clients\Facebook::$autoRefreshAccessToken` is now disabled by default (klimov-paul)
+- Enh #163: Added support for client-specific access tokens at `yii\authclient\clients\Facebook` (klimov-paul)
+- Chg #163: `yii\authclient\clients\Facebook::$autoRefreshAccessToken` is now disabled by default (klimov-paul)
 
 
 2.1.2 February 15, 2017